Loopback odd byte test now works
Odd byte support required changes in both KS firmware and smash command
Fixed handling of more KS data available than will fit in smash buffer
Added Loopback test IOPS measurement
Removed unnecessary use of Amiga ROM overlay (OVL) for flash operations
Reduced KS message reply latency by inlining DMA setup functions

+++ fw/timer.c
@@ -196,7 +196,7 @@ void
 timer_delay_ticks(uint32_t ticks)
 {
     uint32_t start = TIM_CNT(TIM1);
-    while (TIM_CNT(TIM1) - start < ticks) {
+    while ((uint16_t) (TIM_CNT(TIM1) - start) < ticks) {
         /* Empty */
     }
 }
@@ -276,7 +276,7 @@ void
 timer_delay_ticks(uint32_t ticks)
 {
     uint32_t start = TIM_CNT(TIM2);
-    while (TIM_CNT(TIM2) - start < ticks) {
+    while ((uint16_t) (TIM_CNT(TIM2) - start) < ticks) {
         /* Empty */
     }
 }
